#3 stylesの内容を別ファイルに移動
