+++ react-quote-table/src/App.css
@@ -6,6 +6,3 @@
 body{
     font-family: "Nasu-Regular";
 }
-
-
-
