+++ react-quote-table/src/components/molecules/PdfTableHeader/styles.js
@@ -1,0 +1,18 @@
+const styles = {
+  tableHeader: {
+    flexDirection: "row",
+    borderBottom: "1px",
+  },
+  tableHeaderColNo: {
+    width: "10%",
+  },
+  tableHeaderCol: {
+    width: "20%",
+  },
+  tableHeaderCell: {
+    fontSize: "13px",
+    fontFamily: "Nasu-Bold",
+  },
+};
+
+export default styles;
