+++ react-quote-table/src/components/molecules/PdfTableRow/index.jsx
@@ -1,22 +1,6 @@
 import React from "react";
 import { View, Text } from "@react-pdf/renderer";
-
-const styles = {
-  tableRow: {
-    flexDirection: "row",
-  },
-  tableCol: {
-    width: "20%",
-  },
-  tableColNo: {
-    width: "10%",
-  },
-  tableCell: {
-    fontSize: "12px",
-    fontFamily: "Nasu-Regular",
-    margin: "5px",
-  },
-}
+import styles from "./styles.js"
 
 const PdfTableRow = ({ value }) => {
   return (
