+++ react-quote-table/src/components/molecules/PdfTableRow/styles.js
@@ -1,0 +1,18 @@
+const styles = {
+  tableRow: {
+    flexDirection: "row",
+  },
+  tableCol: {
+    width: "20%",
+  },
+  tableColNo: {
+    width: "10%",
+  },
+  tableCell: {
+    fontSize: "12px",
+    fontFamily: "Nasu-Regular",
+    margin: "5px",
+  },
+};
+
+export default styles;
