+++ react-quote-table/src/components/organisms/PdfTableHead/index.jsx
@@ -1,5 +1,5 @@
-import { TableCell, TableHead, TableRow } from "@mui/material";
 import React from "react";
+import { TableCell, TableHead, TableRow } from "@mui/material";
 
 const index = () => {
   return (
