+++ react-quote-table/src/components/organisms/PdfWrapper/styles.js
@@ -1,0 +1,21 @@
+const styles = {
+  wrapper: {
+    margin: "50px",
+  },
+  title: {
+    fontFamily: "Nasu-Bold",
+    fontWeight: "bold",
+    textAlign: "center",
+    marginBottom: "50px",
+  },
+  totalPrice: {
+    fontFamily: "Nasu-Bold",
+    borderBottom: "1px",
+    marginBottom: "50px",
+  },
+  table: {
+    width: "auto",
+  },
+};
+
+export default styles;
