+++ react-quote-table/src/components/templates/Pdf/index.jsx
@@ -1,12 +1,13 @@
 import React from "react";
 import PdfTableWrapper from "../../organisms/PdfWrapper";
-
 import {
   Document,
   Page,
   PDFViewer,
   Font,
 } from "@react-pdf/renderer";
+
+import styles from "./styles"
 
 Font.register({
   family: "Nasu-Regular",
@@ -17,13 +18,6 @@ Font.register({
   family: "Nasu-Bold",
   src: "./fonts/Nasu-Bold.ttf",
 });
-
-const styles = {
-  viewer: {
-    width: window.innerWidth,
-    height: window.innerHeight,
-  },
-};
 
 const index = ({ allInfo, totalPrice }) => {
   return (
