+++ react-quote-table/src/components/templates/Pdf/styles.js
@@ -1,0 +1,8 @@
+const styles = {
+  viewer: {
+    width: window.innerWidth,
+    height: window.innerHeight,
+  },
+};
+
+export default styles;
